% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellForest.R
\name{CellForest}
\alias{CellForest}
\title{CellForest}
\usage{
CellForest(data, k, ncores = -1)
}
\arguments{
\item{data}{A expression data matrix,gene in rows and samples in columns.}

\item{k}{The number of clusters to output.}

\item{ncores}{The number of cores to be used when the program running in parallel.}

}
\value{
samplegeneweight The weights of genes calculated by function randomForest based on random permuted predict label.

geneImportance The weights of genes calculated by function randomForest.

selgenes The genes Cell Forest final select.

selgeneratio The proportion of the number of selected genes in the total genes.

runningtime The running time of the program.

cutoff Threshold for selecting genes.

plabel Predicated clusters.
}
\examples{
data(cfdemo)
result = CellForest(data = data,k = 3)
}
\author{
Hongdong Li, lhdcsu@gmail.com, Central South University.
}
\keyword{clustering,}
\keyword{feature}
\keyword{selection}
\keyword{unsupervised}
