\name{CellForest}
\alias{CellForest}
\title{CellForest}
\usage{
CellForest(data,kcluster,ncores=-1)
}
\arguments{
\item{data}{A expression data matrix,gene in rows and samples in columns.}
\item{kcluster}{The number of clusters to output.}
\item{ncores}{The number of cores to be used when the program running in parallel. When set to -1, the number of cores to use is n-1 where n is the total number of cores.}
}
\value{
\item{geneImportance}{The weights of genes calculated by function randomForest.}
\item{CM}{consensus matrix calculated by averaging all similarity matrices in CMall.}
\item{cluster}{predicated clusters by using CM and Hierarchical Clustering.}
\item{CMall}{consensus matrix for each of the value in toppct.}
\item{clusterall}{predicated clusters for each of the value in toppct.}
\item{randomGrank}{Noise distribution generated by random labels.}
\item{sampleLabelRes}{Statistics of changes in the four thresholds.}
\item{nselgenes}{The number of selected genes.}
\item{selgeneratio}{Proportion of selected genes.}
\item{time}{Program running time.}
}
\description{
Cell Forest is a supervised feature selection-based single cell clustering method, which is built on the hypothesis that using only a subset of cell cluster related genes could improve accuracies for cell clustering .
}
\examples{
data(CFDemo)
result = CellForest(data,kcluster = kprior)
performance = evalcluster(label,result$cluster)
}
\author{
Validation methods implemented by Hongdong Li, lhdcsu@gmail.com, Central South University.
}
\keyword{clustering, unsupervised feature selection}
